Add files via upload
Presently, the major challenge of the medical industry is to predict the cardio vascular disease with less expensive and more reliable method to avoid the compounding effect of the disease in low income or developing countries. The early detection not only reduce the cost but also improves the quality of life.
The aim of this research is to develop an efficient way to predict the presence of the cardiovascular disease. The steps as mentioned below.
(1) The UCI dataset is used to predict the disease.

(2)The Features are selected based on high positive correlation values with the target and used random order of data (without sorting).

(3)The performance of the model is evaluated by Five different training and testing ratio of dataset.

(4)To check the behavior of the model with low to high training and testing data.